requirements.txt
# The single source of truth for the OpenSpiel pip dependencies.
absl-py == 0.7.1
tensorflow-gpu < 1.15.0, >= 1.14.0
dm-sonnet == 1.32
IPython == 5.8.0
tensorflow-probability < 0.8.0, >= 0.7.0
cvxopt == 1.2.3
networkx == 2.2
mock == 3.0.5
matplotlib == 3.1.1
nashpy == 0.0.19
scipy == 1.1.0
attrs == 19.1.0
